Glyph.Index debe ser un int, ya que puede contener una unidad de UTF-32 en vez del índice del glifo.
afortunadamente, el tamaño parece no aumentar.

también se añadió el atributo GlobalClass a MarkerTag.

+++ src/core/rich_text/Glyph.cs
@@ -56,7 +56,7 @@ public readonly struct Glyph // 64 bytes pesa la marranota
 	/// <summary>
 	/// Gets the index of the glyph in the source <see cref="Font"/>, if applicable.
 	/// </summary>
-	public ushort Index { get; private init; } // fonts cannot hold more than 0xFFFF glyphs, so ushort it's fine.
+	public int Index { get; private init; }
 
 	/// <summary>
 	/// Gets the texture resource associated to the icon.
